requirements.txt
Flask
google-generativeai
python-dotenv  # (optional, for environment variables)
